requirements.txt
apturl==0.5.2
bcrypt==3.1.7
beautifulsoup4==4.14.2
blinker==1.4
blis==0.4.1
Brlapi==0.7.0
cachetools==5.5.2
catalogue==1.0.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.11
dbus-python==1.2.16
defer==1.0.6
descartes==1.1.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
en-core-web-sm==2.2.5
entrypoints==0.3
fasteners==0.14.1
filelock==3.16.1
fire==0.7.1
fonttools==4.57.0
fsspec==2025.3.0
ftfy==6.2.3
future==0.18.2
gdown==5.2.0
hf-xet==1.2.0
httplib2==0.14.0
huggingface-hub==0.36.0
idna==2.8
importlib-resources==6.4.5
jinja2==3.1.6
joblib==1.4.2
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
matplotlib==3.7.5
monotonic==1.5
mpmath==1.3.0
murmurhash==1.0.15
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
nuscenes-devkit==1.1.11
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
open-clip-torch==2.32.0
opencv-python==4.12.0.88
packaging==25.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
plac==1.1.3
preshed==3.0.12
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycocotools==2.0.7
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyquaternion==0.9.9
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2024.11.6
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Shapely==1.8.5.post1
simplejson==3.16.0
six==1.14.0
soupsieve==2.7
spacy==2.2.4
srsly==1.0.7
ssh-import-id==5.10
sympy==1.13.3
systemd-python==234
termcolor==2.4.0
thinc==7.4.0
threadpoolctl==3.5.0
timm==1.0.22
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
triton==3.0.0
typing-extensions==4.13.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wasabi==0.10.1
wcwidth==0.2.14
xkit==0.0.0
zipp==3.20.2
